scripts/ingest/requirements.txt
xarray
netCDF4
numpy
pandas
psycopg2-binary
chromadb
sentence-transformers
